// src/app/model/architect-sample.ts
export class ArchitectSample {
  id: string = '';
  architectId: string = '';
  architectName: string = '';

  materialName: string = '';
  category: string = 'Sample';  // Sample | Material
  photoUrl: string = '';
  piecesCount: string = '';
  dateGiven: string = '';
  returned: boolean = false;
  returnedDate: string = '';

  createdAt: number = 0;
}